refactor: split module discovery and loading
