+++ src/dndcs/core/__init__.py
@@ -1,4 +1,4 @@
-from . import registry, models
+from . import registry, models, discovery, loader
 from .module_base import ModuleBase
 
-__all__ = ["ModuleBase", "registry", "models"]
+__all__ = ["ModuleBase", "registry", "models", "discovery", "loader"]
